requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-jet-reboot @ https://github.com/arunmandal53/django-jet-reboot/releases/download/release-v1.3.2/django_jet_reboot-1.3.2-py3-none-any.whl
django-redis==5.2.0
django-reversion==5.0.4
djangorestframework==3.14.0
Pillow==9.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
